OK guys.  I just make a pretty major commit to robotlib which puts simulink in as the back-end (first only for running the plants over LCM, but assuming this works for everyone, I'll change everything else over, too).  All of these changes happened behind the scenes and, if all goes well, the only difference that you should notice is that the plant dynamics when simulated as an LCM node are more accurate now - because they are using proper ODE routines instead of dumb euler integration.  But much more substantially, it allows me to now do the realtime integration of the hybrid models and hybrid controllers (both coming soon) as LCM nodes, and vastly opens up our options for realtime integration routines.
After you update, please do:
>> configure
>> make 
in your main robotlib directory.

In the process of doing this, I did make one MAJOR change to the interface that will effect you all.  Nearly every instance of the word "dynamics" has been changed, as appropriate, to the word "plant" (e.g. runLCMDynamics -> runLCMPlant and PendulumDynamics -> PendulumPlant and so on).  This had to happen because now controllers and estimators will all have dynamics, too, and the dynamics back-end is shared between them.  And it turns out the change really cleans up the readability of the code nicely.  But it might be a pain to change your things over.  I really am sorry, but it's for the best!

Let me know if you have any problems, especially with the new runLCMPlant routine.

- Russ



git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib@103 c9849af7-e679-4ec6-a44e-fc146a885bd3