Got tired of making gradients for every case, so I've added a pretty slick inline autodiff which can do taylor expansions (to arbitrary order) of basically all of our functions.  (might not have thought to implement every case yet, but i can quickly implement any that we trip that are not supported yet).
the current implementation is notably slower than writing the gradients to a file and using that directly... but i'm not convinced that it has to be.  the new version has the advantage that it shares computation of f and df more intelligently.  if i can get the performance up, then it's a serious candidate for replacing every piece of gradient code we have lying around.


git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib@674 c9849af7-e679-4ec6-a44e-fc146a885bd3