resolved sample time issue by disabling feedback / cascade combinations (as drake systems) of ct systems with dt systems.  these will kick out to dynamical systems (simulink models) instead now.
the subtlety is that the sample time of a system also implies a sample time for the output function (in addition to when update and dynamics get called).  so a CT cascaded with a DT system means that the DT systems output function will now be evaluated continuously, even though it was not designed to do that.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@3498 c9849af7-e679-4ec6-a44e-fc146a885bd3