FeedbackSystem now supports more complicated feedback connections (where there are direct feedthrough subsystems of a feedback connection)
separated getInertial method from RigidBody into getMass and getCOM, because I wanted to maintain the derivative convention when returning the jacobian of the com.  (plus scott and kanako seemed to think it was poorly named. :)

fixed COM calculation.  this uses the kinematics code, and doesn't seem to have the same problems that Scott was talking about in his fix.

added COM visualization to PlanarRigidBodyVisualizer when the debug property is set to true.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@4592 c9849af7-e679-4ec6-a44e-fc146a885bd3