implemented the mex_model_ptr in the right way.  (short explanation below)
previously, matlab was acting as if mex_model_ptr is passed by value, when it was actually being passed by reference.  e.g., if I had

p = RigidBodyManipulator(...)

p2 = setInputFrame(p,...);
p2.compile();

p.simulate()

I was going to have a problem, because p2.compile will delete (in cpp) the mex model ptr that p was using, but p doesn't have a clue that this happened.

i implemented a solution here, which unfortunately required a reorg of the files...
i created the class SharedDataHandle that could be used like

obj.mex_model_ptr = createMexModelPtr(HandCmex(obj),@deletefcn);

then using obj.mex_model_ptr.getData to pass things into the mex.  this way I can get away with never calling the delete function explicitly, and instead rely on the "garbage collection" mechanism of the handle class to free the memory at the right time.

the only problem is that my external handle class could not call the mex freeing function because it is a private method of the RigidBodyManipulator.  so to make it work, I'm had to undo some of the cleaning up that I did before (tucking the mex files into the classes).

but it all works well now.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@4878 c9849af7-e679-4ec6-a44e-fc146a885bd3