ik mex runs on my laptop.  looks like it still gets different answers sometimes.
it used to run the full atlas ik in ~.25 seconds.
i found a stupid error (it wasn't actually using the gradients i was providing in matlab) that got it down to ~.1 seconds.

now i've mex'd the snopt part, too, and it's down to ~ .02 seconds.   that is with the default convergence tolerances (which are quite tight).

i think that there are a lot of places where we could be more efficient with memory allocation in the kinematics mex code.  (eigen makes it too easy to dynamically resize things all over the place).  i think we can probably make all of the kinematics code faster by combing through that.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@5361 c9849af7-e679-4ec6-a44e-fc146a885bd3