resolves bug 1613: matlab crashes on megaclear
Short story:  i was trying to be clean/clever with my new drakeMexPointer class and support the calling of destructors in c (without having to explicitly make a new outward facing mex interface).  that turned out to have a problem - it was possible for the mex function (which owned the memory with the method implementation) to get cleared before the mex pointer destructor was called.  --> segfault.  So I've changed the DrakeMexPointer to instead take the name of a method to call when it's time to delete the pointer, and restored the deleteModelmex.cpp to it's rightful place.

This is a little more limiting than what I had before, where I could take any function handle, and that function handle was constructed in matlab.  Now that the mex pointers are being constructed in C, it's not quite so easy.  If we want to support other arguments being passed in to the delete function (e.g. so that the one could have a mex function mymex which takes mymex(1) to build the pointer and mymex(2,ptr) to destroy the pointer), then we can add that functionality in as it is needed.  Just let me know.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@6701 c9849af7-e679-4ec6-a44e-fc146a885bd3