The systemWExports function worked well, but had one major problem : the order of the paths in LD_LIBRARY_PATH matters, and I could not back that out using ldd / otool.  So I've zapped that method.
The replacement approach is to write the LD_LIBRARY_PATH to the CMakeCache during the build, load it and execute using it with a new method systemWCMakeEnv.  That seems to be a much more robust solution.  I should consider saving other environment attributes, like PKG_CONFIG_PATH, etc and adding them to this method, too.  But I'd prefer to do it as necessary instead of trying to make an excessive exhaustive list now.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@6736 c9849af7-e679-4ec6-a44e-fc146a885bd3