faster creation of UID from time (from Robin)
note from Robin:

I've been playing around with speeding up the walking planning system, and most
of the time spent on planning is in the construction of DrakeSystems (mostly
PPTrajectories). A significant chunk of that time is just spent on the
generation of the UUID using datestr(). The attached patch is over 20 times
faster and does not break any Drake or DRC unit tests (in fact, it should
result in a "more" universal ID, since it uses the system clock down to the
microsecond, rather than millisecond).

sprintf seems to be by far the fastest way to do this:

========
>> tic; for j = 1:100; u = datestr(now, 'MMSSFFF'); end; toc
Elapsed time is 0.134840 seconds.
>> tic; for j = 1:100; u = num2str(now * 24*60*60*1e6, '%018.0f'); end; toc
Elapsed time is 0.028680 seconds.
>> tic; for j = 1:100; u = sprintf('%018.0f', now * 24*60*60*1e6); end; toc
Elapsed time is 0.005879 seconds.
>>
========

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7139 c9849af7-e679-4ec6-a44e-fc146a885bd3