patch from hongkai:
We can check samples in between knot points for inverseKinTraj, this will be very useful in collision avoidance, as the previous version only checks collision avoidance at the knot points
Now MultipleTimeKinematicConstraint supports evaluating time breaks that are not all within tspan. The previous version only handles the case that all time are within tspan, otherwise it would just quit evaluating the constraints.
Based on Scott's suggestion, I set the robotnum argument of WorldCoMConstraint and QuasiStaticConstraint to be optional, instead of a compulsory argument.
Fix a bug in the cubic spline computation in inverseKinTraj. I think this is probably the reason why inverseKinTraj sometimes threw INFO 41 before.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7575 c9849af7-e679-4ec6-a44e-fc146a885bd3