prettier printing for coverage.  handled known bug in executed lines (see transcript w/ mathworks contact Phil below)
-----
Yes you will need to handle that separately.  It is a known bug that will be fixed in a future version of MATLAB.  Did you ever test/try the LXE?  It is fixed there.

Phil

-----Original Message-----
To: Philip Borghesani
Subject: Re: command line access to the code coverage tool

Phil,

I've got the basic functionality up and running now.  Thanks for the leads!

One minor annoyance is that callstats('file-lines',filename) seems to return the classdef lines as executable lines, but they are never called.  Am I correct about that (I should handle it separately) or is there something I'm missing?

- Russ

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7639 c9849af7-e679-4ec6-a44e-fc146a885bd3