patch from Robin resolving bug 1879.
bad distance function from current PolygonalObstacle2D PolygonalObstacle2D correctly checks for the edge case of vertical obstacle boundaries, but it doesn't get horizontal edges right. The attached figure shows the value of phi evaluated over a grid, from an obstacle with corners at [0,0], [0,1], [1,1], [1,0]. Those weird bulges and edges along the y=0 boundary are the result of this bug. Rather than adding another edge case, I've rewritten polyconstraint in terms of the a'*[x;y] = b formulation rather than slope-intercept, which has no special cases for horizontal or vertical lines. I've also added a test case (testVerticalAndHorizontalObstacles.m) which checks distances from horizontal and vertical edges. This case fails (as expected) on the trunk version of PolygonalObstacle2D and passes on the new version. I've also run the testObstacles test case and checked every value returned by the old polyconstraint against the new one, and they are the same in every case (testObstacles uses a fixed obstacle field which happens to contain no horizontal edges, so the problem doesn't show up in that test). The new polyconstraint is about 50% slower, but it's cleaner and more correct, so I think that's a win overall.

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7743 c9849af7-e679-4ec6-a44e-fc146a885bd3