after making (and then locally reverting) a relatively big change, I’ve now just added implementation note:
for homogenous transforms, I could do the following
vector transforms more efficiently.  forwardKin is adding a 1 on
the end of every pt for the homogenous coordinates.  it would be
equivalent, cleaner, and more efficient, to just add a zero on the
end instead of the 1...  because for homogeneous transform matrix
T we have:
   T * [x;1] - T * [0;1] = T * [x;0]
but I made this change everywhere and decided it was more important
to keep the interface to the forwardKin method clean instead of
polluting it (and bodyKin, and the mex files, ...) with an extra
input/option for "vectors_not_points".

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7967 c9849af7-e679-4ec6-a44e-fc146a885bd3