Russ,
See attachment for a patch that adds a geometricJacobian method to RigidBodyManipulator. I made the patch in the directory drake.

function [J, vIndices] = geometricJacobian(obj, kinsol, base, endEffector, expressedIn)
%GEOMETRICJACOBIAN Computes the geometric Jacobian from base to endEffector
% expressed in frame attached to expressedIn
%   The general contract of this method is that for joint velocity vector
%   v, the twist of endEffector with respect to base, expressed in
%   expressedIn, can be computed as J * (v(vIndices)).

I decided to add the option to specify the frame in which the Jacobian is expressed since it was so easy to add. I was mistaken about the order of omega and v in Featherstone's algorithms, it is indeed [omega; v], so I adopted that as well.

The geometricJacobian method allows you to compute the Jacobian between any two bodies in the kinematic tree, and takes sign changes due to going up and down the tree into account.

I added a test that compares twist = J * (v(vIndices)) to the relative Tdot between random bodies for Atlas, as computed by doKinematics.

One caveat: it doesn't work for floating joints as they are currently implemented. If you parameterize the velocity across a floating joint simply as the twist across the joint in body frame, the motion subspace is simply identity. Instead, if you use yawd/pitchd/rolld or quatd for rotation, it's going to be configuration dependent. I can incorporate that too, but maybe I shouldn't bother if we're going to switch to using twist as the parameterization of velocity across a floating joint.

- Twan

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7993 c9849af7-e679-4ec6-a44e-fc146a885bd3