Fix #13 by syncing frames within Atlas.
The problem was that `pdcontrol` looks for the state frame of the
`RigidBodyManipulator` in the output frame of the
`TimeSteppingRigidBodyManipulator`. Since `Atlas` resets it's state, input, and
output frames to Atlas specific versions, that lookup failed. I've resolved
this by having the `Atlas` constructor reset the state, input, and output
frames of the `RigidBodyManipulator` as well.

Note: This setup will need some tweaking when we want to add sensors to `Atlas`
objects. The approach we've used up to this point (resetting frames with
Atlas specific ones) will clobber the changes made to those frames by the
addition of the sensor. I'll create an issue for this.