Sanitize `state_dims` before use as loop variable
The overloaded `inspector` method in `RigidBodyWRLVisualizer` was
passing `state_dims` to the parent method as a row vector. The method in
`Visualizer` was then transposing that before using it as the loop
variable in a for loop. Since for loops loop column-wise, I've replaced
the transpose with a reshape to ensure that the variable passed to the
for loop is a row vector. Fixes #23.