Get bullet gradients working (messy commit)
Just want to make sure I don't lose the progress I've made here. The most
important change, however is not in Drake - Bullet ought to be compiled with
double precision (which is *not* the default). There's a switch in the Bullet
CMakeLists.txt. Any code that includes the Bullet headers also needs to contain

```
```
or that flag needs to be set for the whole build.