Return one potential contact per geometry pair
This is a change from the previous approach, which returned one potential
contact per pair of bodies. The old approach could lead to the closest points
jumping from one geometry to another. The downside to the new approach is that
there are many more potential contacts for systems with many contact shapes per
body.