Make `PR2/{runPassive,runSagitalPassive}` run
This required changing these to use `TimeSteppingRigidBodyManipulator`s with
`ignore_self_collisions` set to `true`. If they were left with regular
`RigidBodyManipultor`s they failed while trying to resolve constraints in
`DrakeSystem/getInitialState`. This occurred even with the ignore
self-collisions flag set because the plants were generated as fixed base
manipulators and therefore the PR2 body was part of the world. Several links
are embedded inside the body link which are not children of that link. This led
to unresolveable collisions.