Load meshes in `URDFRigidBodyManipulator`
Also fixes handling of collision elements so that the collision elements of
links with no "default" collision group are still parsed. Adds
`testURDFcollision` to check that `collisionDetect` returns the same
information for `RigidBodyManipulator` and `URDFRigidBodyManipulator`. Note,
the order of points returned, as well as the body indices are *not* the same
between these two classes. However, all the same collision pairs are checked.