Make warnOnce warn once per instance.
The disabled_msgids map was being default initialized in the classdef of
WarningManager. Because containers.Map is a handle class, this means that all
WarningManager objects shared the same disabled_msgids map. So rather than
warning once per instance of a class containing a WarningManager, we were only
warning once ever. The fix is to initialize the disabled_msgids in the
constructor instead.