Hand control demo functional
Fixed visualization bug that was keeping hand state from
being drawn, and added hand-weight-simulator s.t. the
controllers know about the extra force on the wrist
from the hand. The weight I'm adding needs to be updated
to be precise... I just guessed, but it can be calculated
from the urdf