move atlas frames and controllers into packages
This makes it possible to unambiguously refer to them. Previously, there
were name clashes between the DRC frames/controllers and the Drake
frames/controllers, which meant that switching back and forth between
the two without restarting matlab was difficult or impossible.

As an additional bonus, by putting the frames and controllers into
packages, we no longer need to do any addpathTemporary() calls to access
those files; in fact, we don't need to make any modifications at all to
the user's path.