Fix case where parent and child belong to the CFG
We weren't checking whether the parent belonged to the collision filter group
until too late. The result was that if both parent and child had collision
geometry, the welded link would never belong to the collision filter group -
even if both parent and child did. By checking wether the parent belongs to the
group first, we can avoid this problem.