explicitly consider the fact that the LIP height changes when walking
We've been constructing the linear inverted pendulum ZMP solutions by
either hard-coding a pelvis height above the feet, or using the current
CoM height. Neither of these is exactly right. When standing, we should
account for the fact that we will crouch down when we start walking. But
when doing a recovery plan, we won't necessarily crouch, so we should be
able to use the current height instead.