Add exp2quat matlab functions
exp2quatImpl.m is set up for codegen to produce mex or lib. exp2quat.m is set
up to use mex if available and fall back to MATLAB if not. exp2quatTest tests
first and second derivatives with TaylorVar for the non-degenerate case and
first derivatives numerially for the degenerate case.