btCollisionObjects w/o a user ptr can collide
We set the user pointer on all btCollisionObjects that are part of the
collision model, but when evaluating point-cloud to model collisions, we create
new btCollisionObjects which do not have a set user pointer because they don't
belong to a DrakeCollision::Element. These should still be able to collide with
the model.