restore matlab/c++ duplication for CoordinateFrame implementation.  for now...
related to #1267 .  i am hereby removing the stub interface files (which I believe to be correct, but which cause unexplained behavior in matlab... as described in the pull request #1274).

only change vs master should be changing the getaccess for the coordinates member to private.  that will make our lives better soon.
i'm hoping this will let me get past the build servers so I can merge.