Add a simple "lidar simulator"
Add a simple test program to publish lidar messages. This generates
random point clouds that fit into a semblance of a hypothetical 360°
volume that a lidar sensor might scan and publishes them at a rate of
about 10 Hz. This isn't very realistic, but should suffice for simple
tests of the message format and ability of consumers to receive and
unpack the data.