Optimization: Separate variable bindings from constraints
For each `Constraint` added to an `OptimizationProblem` we need to store
variable bindings that map decision variables from the problem onto
input variables to the constraint function.  Delay this association
until a constraint is added to a problem, and store it inside the
problem.  This avoids storing anything specific to one problem in a
constraint, and will allow a constraint to be shared among multiple
problems simultaneously.

GitHub-Issue: RobotLocomotion/drake#1666