Update std::shared_ptr argument passing to be closer to idiomatic C++.
The CppCoreGuidelines recommend that std::shared_ptr only be passed as
an argument when the callee explicitly needs to manage the lifetime of
the argument, and that 'const std::shared_ptr<T>&' should only be used
if there is a chance that lifetime will not be shared.

https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Rr-smartptrparam

Many instances in drake were passing 'const std::shared_ptr<T>&'
around, often when no lifetime management was necessary, and otherwise
when lifetime sharing was unconditional.  This can be less efficient,
unnecessarily restrictive, and hides the intent of the API.

This patch converts those instances either to:
 * [const] T& if no lifetime sharing is required
 * std::shared_ptr<T> if lifetime sharing is always required