Actually respect CMake build directory
Modify external project build rules to use the build directory as the
parent for external project build directories, rather than building
everything in the source directory. This makes the build more conforming
with CMake standard practice and user expectations when invoking CMake
directly. (Also includes updates to some externals to pull recent
changes to the same to work as "CMake POD's".)

Note that the old (POD-conforming) behavior is retained when using the
Makefile.