Added FCL to build and added FCLModel
Drake now builds with External/fcl as an option. It is off by default
because fcl uses boost which conflicts with the Matlab boost.
FCLModel can be constructed instead of BulletModel. BulletModel is
still the default when newModel() is called.

.../collision/test/FCLModelTest.cpp excersizes some of the tests in
.../collision/test/ModelTest.cpp and compares the results of BulletModel
and FCLModel with the same calls.

I did not implement fcl collision with mesh shapes because the
API appears to only have access to points and not triangles.

I did not implement margins or ray intersections because I could
not find these features in fcl.

With tests or examples of features, I could implement more.