Rephrase some tricky expressions for clarity.
This was taking a `MatrixXd* Q`, then calling `->operator()` and
passing in some index (or indices), which returned a `double&`,
whose address was then taken and passed to a C function that
writes through that `double*` pointer back into the matrix.

This reasonably confused cpplint's runtime/casting rule (who thought
we were taking the address of a temporary), and well as likely
confusing most other readers, including myself.

I believe this rephrasing should be safe, in that if the `operator()`
return type ever changed, the compiler would reject the assignment to
a non-const reference.