"Fix" default value of install prefix for non-pods
Disable the complex logic that tries to locate a special "build" folder
to set a default CMAKE_INSTALL_PREFIX when doing a non-pods build, and
instead use `${CMAKE_BINARY_DIR}/install` as the default. (Also drop the
unnecessary logic to create the directory; CMake will handle this.) The
"detection" logic has some problems and sets a superfluous cache
variable.

It's still interesting to override the default prefix because it is used
during the build to install sub-packages, so the default may not be
appropriate (especially on UNIX systems where the default /usr/local is
usually not user-writable).

Fixes #2093.