Optionally keep external source in build directory
Add an option (only available in non-PODS configuration) to keep the
sources for external projects in the build directory, rather than the
source directory. This ensures that it is always possible to have
multiple build trees for a single (drake) source tree, at the expense of
added inconvenience and disk usage (because the sources for external
projects are duplicated).