cmake: Simplify infrastructure for Git cloning into source directory
We no longer need to distinguish the `${proj}_SOURCE_DIR` and the
`${proj}_GIT_CLONE_DIR` because they are now the same for all cases.  It
is important to keep these directories the same so that a single
`ExternalProject_Add(${proj} ...)` call can manage all external project
steps.  This will later allow us to refactor the `download-all` target
to avoid needing a separate `ExternalProject_Add(download-${proj} ...)`
call.