My original definition of NAryState followed the example set by Drake::CombinedVector, hence it was a template taking two parameters:   a ScalarType, and the *template* of the contained type (which would   become specialized on ScalarType).
Something about how this interacted with NArySystem turned into something
that only g++ was happy to compile; clang said no.  So... now NAryState
is a template on only one parameter, the concrete UnitVector type (an
actual type, not a template).  It does some finagling via decltype() to
recover the 'ScalarType' from UnitVector.  clang is happy.