director: Do not make a directory inside the source tree
The CMake ExternalProject module produces a `mkdir` step that makes
relevant directories before other steps, including the build directory.
For `director` we set the BINARY_DIR to be inside the source tree
because that is where we want `make` to run.  When we later consolidate
our `download-${proj}` targets into their main `${proj}` targets then
this will cause a subdirectory to be made in the director source tree
before its `download` step.  This will cause `git clone` to fail
complaining that the clone destination directory is a non-empty
directory.  Avoid this problem by adding a `chdir` step to our in-source
`director` BUILD_COMMAND so that we can use a BINARY_DIR that does not
sit in a subdirectory of the source tree.