Refactor drake CXXFLAGS
Don't unilaterally promote warnings to errors, as this potentially
breaks the build for any user that has specified additional warnings
(which become errors); instead, explicitly set to errors those warnings
that we want to always enforce. Also, refactor CXXFLAGS so that GCC and
Clang share the same logic (which is sensible as they generally support
the same warnings). This has the further effect of turning on some
further warnings-as-errors for both compilers (which previously were
enabled for only one of the compilers).