Fix strange constructor error
Rewrite member initializers to use explicit construction and assignment
(which is also more understandable) rather than more obscure direct
member construction. This addresses an issue where at least some
versions of GCC complain that the other form would use an explicit
constructor. (This may be a compiler bug, but the generated code should
be the same, and the new version is anyway more readable, so there is
little penalty to writing these in a way to avoid the issue.)