Delete Model copy/assignment
Make copy constructors and assignment operators for Model and subclasses
explicitly deleted, rather than providing implementations(!). Making
them private and NOT implemented is a common pre-C++11 pattern, but
since we require C++11, explicitly deleting them is much better. The
implementations are almost certainly broken, in case they could ever be
called. This also avoids warnings about not explicitly initializing the
base class in the constructor (which can't be done readily, as the base
is not copyable...).