Don't use shared_ptr in fcl_model_test
Use unique_ptr and pass a reference rather than shared_ptr (with an
egregious copy) to pass a model to test cases in fcl_model_test. There
was really no reason to use a shared_ptr here. Also, use the new form of
newModel(TYPE) to create the models, for consistency.