Build "documentation" target only on explicit request
Remove the `documentation` target from the default build targets so that
developers that enable `BUILD_DOCUMENTATION` do not have to wait for the
documentation to re-build every time they build.  Instead it will build
only through the `make documentation` step that we already document.

Add an undocumented option to allow CI builds to drive documentation
as part of the default build.