Fix submodule download in pod-build mode
When using the PODS Makefile for an in-source build, the `POD_BUILD`
option to CMake causes `pod-build` directories to be created in the
externals' source trees.  The `pod-build` directory prevents the
`git clone` performed by `git submodule update` during the download
step from seeing an empty directory, so it refuses to clone.  Teach
the download step in `POD_BUILD` mode to remove the `pod-build`
directory prior to cloning.

Fixes #2565.