Fix thread linking
Create an interface target to simplify linking to the platform thread
library (if any). Modify several exexuctables to link to this.

This is required as the (template, inline) function `runLCM`, which
uses `std::thread`, apparently requires linking to the pthread library
on some platforms. This previously worked because using LCM via
pkg-config carried along a number of not-always-necessary libraries,
including pthread, but the new exported targets are much more
conservative.