Disable clang sanitization in lcmgen
Update LCM to a branch that disables clang {address,memory} sanitization
when building lcm-gen. This is needed as lcm-gen is sloppy about its
memory management (reasonably so, as non-systemic leaks in a short lived
process aren't really a problem), and so building with sanitization
results in a broken executable, which causes problems later when we need
to generate LCM bindings.

This also fixes a build error on Ubuntu if Lua is installed. (If LCM
finds Lua, it tries to build the Lua module by default, but fails due to
the headers on Ubuntu being in a non-standard location and because the
include directory was not being added.) This was fixed upstream.

Upstream PR (pending): https://github.com/lcm-proj/lcm/pull/78