Fix sign compare warnings from issue #2743.
This patch just makes the type of the loop counter variables match the
type being compared to.  While size_t is generally not favored in
Drake code, this is a MATLAB interface, and the function mxGetN() is
documented as returning
size_t. http://www.mathworks.com/help/matlab/apiref/mxgetn.html