Fix build commands with requirements
Modify how drake_add_external processes parses user specified commands
to allow delayed expansion of variables. This is required for variables
which are defines by an external's required packages, as immediate
expansion may expand to nothing.

In particular, this fixes the build of textbook. (For some reason, it
only seems to break with makefiles, even though it ought to be broken
everywhere.)