Fix MSVC version error
Fix the error message when MSVC is too old; the old error message
claimed to require "14", when it actually requires 19 (a.k.a. VS 2015).
Also, note the Visual Studio release year in the message, which is more
likely to be recognized by users than the internal monotonic version
number. Also, streamline the logic for performing the version check and
spitting out an error if the compiler version is not adequate.