cmake: Sync submodule URL changes during build
When `AUTO_UPDATE_EXTERNALS` is `ON` we are responsible for keeping
external source trees up to date.  Previously we achieved this by
running `git submodule sync` during CMake configuration to honor changes
to the submodule URLs.  However, it is possible for an incremental
update to re-build without re-running CMake.  Therefore we need to do
the sync at the start of the build instead.

We need to add dependencies very carefully to make sure this works with
the CMake Ninja generator.  Any target to which CMake may propagate the
download step custom command of an external project must explicitly
depend on the submodule-sync target.  This is due to peculiarities of
how the Ninja generator consolidates dependencies of custom commands
that propagate to multiple targets.

Fixes #2827.