Refactor external project logic
Consolidate and refactor the logic to set up external projects (and
external examples) in the drake superbuild, moving it to functions in a
separate file. Add new options `ALWAYS` and `LOCAL`, indicating an
external that is always built and/or is local to the repository (does
not need download or update steps), which together allow us to eliminate
logic that special-cased certain externals (notably drake itself).

In addition to the general cleanup, this changes the format for setting
up an external from mostly-declarative to mostly-procedural, which is
less repetitive, better organized, and should be easier to read.

The documentation of the newly refactored build logic is also improved.