Fix inclusion of textbook
Fix check whether we want to build the textbook. The old check asks if
the directory exists, but since we switched to submodules, the directory
will ALWAYS exist. Instead, check for a CMakeLists.txt.

Presumably, the only reason this is not currently causing build failures
when the textbook is not enabled is because `if(IS_DIRECTORY)` "is
well-defined only for full paths", and is likely evaluating to false,
even when we actually DO want the textbook.