Address libbot / bot_core_lcmtypes conflict
Both libbot and bot_core_lcmtypes generate bindings for the libbot LCM
types. Since this means that there is significant overlap in the output
files, including the C library, this can result in odd, apparently
random errors if the two externals are built in parallel. Address this
by adding an artificial dependency between the two to force them to be
built in serial.

(In the long run, we hope to split lcmgl from libbot, and build only
that and the stand-alone types package, dropping the monolithic libbot
package entirely.)

Fixes #2865, hopefully.