Refactor superbuild options
Move superbuild options into their own file, and add helpers for setting
up options. Improve some option descriptions. Use `USE_SYSTEM_<name>`
for options whether to build our own copy of a required package, as this
is a common naming convention for such, and makes it clear which
packages are required and which are optional.